+++ histou/bootstrap.php
@@ -6,7 +6,7 @@ PHP version 5
 @package default
 @author Philip Griesbacher <griesbacher@consol.de>
 @license http://opensource.org/licenses/gpl-license.php GNU Public License
-@link https://github.com/Griesbacher/histou
+@link https://github.com/ConSol/histou
 **/
 
 function getPath()
@@ -14,10 +14,12 @@ function getPath()
     return substr(realpath(dirname(__FILE__)), 0, -6);
 }
 
-function __autoload($className)
-{
+spl_autoload_register('histou_autoload');
+
+function histou_autoload($className) {
     $file = strtolower(str_replace('\\', DIRECTORY_SEPARATOR, $className)).'.php';
     if (file_exists($file)) {
         require_once $file;
     }
 }
+