+++ histou/grafana/graphpanel/graphpanelfactory.php
@@ -6,7 +6,7 @@ PHP version 5
 @package Histou
 @author Philip Griesbacher <griesbacher@consol.de>
 @license http://opensource.org/licenses/gpl-license.php GNU Public License
-@link https://github.com/Griesbacher/histou
+@link https://github.com/ConSol/histou
 **/
 namespace histou\grafana\graphpanel;
 
@@ -17,7 +17,7 @@ PHP version 5
 @package Histou
 @author Philip Griesbacher <griesbacher@consol.de>
 @license http://opensource.org/licenses/gpl-license.php GNU Public License
-@link https://github.com/Griesbacher/histou
+@link https://github.com/ConSol/histou
 **/
 class GraphPanelFactory
 {
@@ -29,8 +29,8 @@ class GraphPanelFactory
     {
         if (DATABASE_TYPE == INFLUXDB) {
             return new \histou\grafana\graphpanel\GraphPanelInfluxdb($title, $legendShow, $id);
-        } elseif (DATABASE_TYPE == ELASTICSEARCH) {
-            return new \histou\grafana\graphpanel\GraphPanelElasticsearch($title, $legendShow, $id);
+        } elseif (DATABASE_TYPE == VICTORIAMETRICS) {
+            return new \histou\grafana\graphpanel\GraphPanelVictoriametrics($title, $legendShow, $id);
         } else {
             throw new \InvalidArgumentException("The given Database is unkown:".DATABASE_TYPE);
         }